// packages/base/plugin-system/src/types.ts
export interface PluginDefinition {
  /**
   * The id of the plugin.
   */
  id: string;

  /**
   * The plugin provider path/package name that provides this plugin.
   */
  provider: string;

  /**
   * The name of the plugin.
   */
  name: string;

  /**
   * The version of the plugin.
   */
  version: string;

  /**
   * A description of the plugin.
   */
  description?: string;

  /**
   * The plugin's base path/path to the package.json file
   */
  packagePath: string;

  /**
   * The path to the plugin's configuration file.
   */
  configPath?: string;

  /**
   * An optional path to an image to use as the icon for the plugin.
   */
  imagePath?: string;

  /**
   * A list of plugin providers that this plugin depends on.
   */
  dependencies: string[];

  /**
   * A set of optional tags that describe the plugin.
   */
  tags: string[];

  /**
   * The options to pass to the plugin module when instantiating
   */
  options: any;

  /**
   * The path to the plugin's module loader.
   */
  loader: string;
}

/**
 * A function that can be used to hook into the plugin system.
 */
export type PluginHookFn<TContext = any> = (
  params: TContext
) => MaybePromise<TContext | ((params: TContext) => MaybePromise<TContext>)>;

/**
 * A plugin module that can be loaded by the plugin system.
 */
export interface IPluginModule<TContext = any> {
  hooks?: Record<string, PluginHookFn<TContext>>;
}

/**
 * A plugin loader that can be used to load a plugin module.
 */
export interface IPluginLoader<
  TContext = any,
  TPluginModule extends IPluginModule<TContext> = IPluginModule<TContext>
> {
  load: (
    definition: PluginDefinition,
    options: Record<string, any>
  ) => Promise<PluginInstance<TContext, TPluginModule>>;
  isValid: (module: TPluginModule) => boolean;
  process: (
    context: TContext,
    instance: PluginInstance,
    options: Record<string, any>
  ) => Promise<void>;
}

/**
 * An instance of a plugin that has been loaded by the plugin system.
 */
export interface PluginInstance<
  TContext = any,
  TPluginModule extends IPluginModule<TContext> = any
> {
  loader: IPluginLoader<TContext, TPluginModule>;
  definition: PluginDefinition;
  module: TPluginModule;
  options: any;
  resolvedPath: string;
  executionDateTime: StormDateTime;
}